refactor: registry structure and update ESLint configuration for node imports
