+++ package.json
@@ -20,6 +20,7 @@
     "dotenv": "^16.4.7",
     "eslint": "^9.0.0",
     "eslint-config-prettier": "^9.1.0",
+    "eslint-plugin-import": "^2.31.0",
     "eslint-plugin-yml": "^1.14.0",
     "husky": "^9.0.11",
     "lint-staged": "^15.2.2",
@@ -37,7 +38,7 @@
     ".": "./dist/index.js",
     "./chains/schema.json": "./dist/chains/schema.json",
     "./chains/*": "./dist/chains/*",
-    "./fs": "./dist/index-fs.js"
+    "./fs": "./dist/fs/index.js"
   },
   "types": "./dist/index.d.ts",
   "typesVersions": {
