+++ src/fs/index.ts
@@ -1,0 +1,3 @@
+// Exports for utilities that require fs access and are not suitable for browser use
+export { FileSystemRegistry } from './FileSystemRegistry.js';
+export { getRegistry } from './registry-utils.js';
