--- src/index-fs.ts
@@ -1,3 +1,0 @@
-// Exports for utilities that require fs access and are not suitable for browser use
-export { FileSystemRegistry } from './registry/FileSystemRegistry.js';
-export { getRegistry } from './registry/registry-utils.js';
